Add support for quit, resize messages
In Windows a user can click on the close window button.
We need to notify the view client that close was requested so
it can initiate graceful shutdown of the application.

I started adding support to specify the window style as well, however,
there are some challenges to figure out how to make that work.